implement an example tarpc cli client
- extensive refactoring across the workspace with addition of feature flags.
- remove unnecessary diesel posgtres inclusion in the dependency tree, that causes compilation failures.
- suppress clippy warnings.
