+++ crates/imbibe-macros/Cargo.toml
@@ -1,7 +1,8 @@
 [package]
 name = "imbibe-macros"
 version.workspace = true
-edition = "2024"
+edition.workspace = true
+readme = "README.md"
 repository.workspace = true
 publish.workspace = true
