+++ crates/imbibe-persistence/src/lib.rs
@@ -3,5 +3,6 @@ pub mod store;
 
 mod record;
 
+#[cfg(feature = "database")]
 #[rustfmt::skip]
 mod schema;
