+++ crates/imbibe-persistence/src/store/error.rs
@@ -1,7 +1,5 @@
 use crate::record::error::InvalidValueError;
 
-pub type Result<T, E = StoreError> = core::result::Result<T, E>;
-
 #[derive(Debug, thiserror::Error)]
 pub enum StoreError {
 	#[error("db error: {0}")]
