+++ crates/imbibe-protos/Cargo.toml
@@ -2,7 +2,7 @@
 name = "imbibe-protos"
 version.workspace = true
 edition.workspace = true
-readme.workspace = true
+readme = "README.md"
 repository.workspace = true
 publish.workspace = true
