+++ crates/imbibe-querier/Cargo.toml
@@ -2,6 +2,7 @@
 name = "imbibe-querier"
 version.workspace = true
 edition.workspace = true
+readme = "README.md"
 repository.workspace = true
 publish.workspace = true
 
@@ -10,6 +11,7 @@ workspace = true
 
 [features]
 default = []
+server = ["imbibe-persistence/database"]
 tarpc = [
   "dep:serde",
   "dep:tarpc",
