+++ crates/imbibe-querier/src/error.rs
@@ -1,7 +1,5 @@
 use imbibe_persistence::{pool::PoolError, store::StoreError};
 
-pub type Result<T, E = QuerierError> = core::result::Result<T, E>;
-
 #[derive(Debug, thiserror::Error)]
 pub enum QuerierError {
 	#[error("store error: {0}")]
