+++ crates/imbibe-telemetry/Cargo.toml
@@ -2,6 +2,7 @@
 name = "imbibe-telemetry"
 version.workspace = true
 edition.workspace = true
+readme = "README.md"
 repository.workspace = true
 publish.workspace = true
